fuzz oracle: bind reference decoder by name + premise rework; enforce 5 conformance clauses the corrected oracle surfaced
Scheduled-Fuzz was red since 2026-07-12: ffmpeg_oracle_decode panicked at
the reject-parity arm on every input our decoder legitimately decoded.
Root cause was the harness itself — the oracle bound its decoder with a
wrong numeric AVCodecID, so the whole differential run compared H.264
output against a decoder for a different codec that rejects essentially
everything.

Harness rework (fuzz/):
- bind by avcodec_find_decoder_by_name("h264"); runtime premise guard
  reads the documented AVCodec.name field and fails loudly on mismatch
- open with err_detect=explode + threads=1: oracle frames are never
  concealment output, errors surface deterministically, and a hostile
  SPS can no longer multiply giant frame allocations across thread
  contexts (phony rss-limit OOMs)
- classify our decode before parity asserts: Concealed
  (decode_error_count > 0) and Incomparable (>8-bit / monochrome /
  separate-plane SPS) outputs are exempt; clean frames must match the
  oracle frame against ANY drained frame (single-packet feeding models
  drain differently); "reference rejected, we decoded cleanly" is now an
  advisory [oracle strictness] stderr report — the reference layers
  stream-discipline heuristics beyond bitstream conformance (it refuses
  non-paired fields, which §7.4.1.2.4 / Annex C permit), so a hard
  assert there reds the schedule on every such corner
- SPS-area pre-gate (64x64 MBs) keeps monster declarations out of the
  differential loop; macOS Homebrew dylib paths in the loader candidates
  so the oracle reproduces locally; examples/repro_oracle.rs triages a
  crash artifact in one command

Decode-side conformance gaps the corrected oracle then surfaced, all
now refused as invalid data with the clause cited:
- §8.3.1.2 / §8.3.2.2 / §8.3.3 / §8.3.4 — intra prediction modes whose
  required neighbour samples are unavailable ("shall be used only
  when ... marked as available")
- §8.2.4 — P/SP/B slice with an empty reference DPB (RefPicList0 would
  be all 'no reference picture')
- §8.2.5.2 — motion compensation against a synthesised "non-existing"
  frame_num-gap placeholder reference
- §7.4.3 — frame_num neither PrevRefFrameNum nor (PrevRefFrameNum+1) %
  MaxFrameNum while gaps_in_frame_num_value_allowed_flag is 0
- NAL types 2/3/4 (slice data partitions, unimplemented) surface an
  error instead of silently dropping coded slice content

New #[doc(hidden)] diagnostics on H264CodecDecoder — decode_error_count()
/ active_sps() / stored_sps() — let the harness (and any caller) tell
clean output from partial/concealed output. Regression suite
tests/fuzz_oracle_regressions.rs pins the classification of all eight
archived fuzzer inputs (tests/fixtures/fuzz_regressions/) plus
zero-error decode of a conforming fixture.

Verified: full test suite green (50 binaries, 30/31 corpus targets
BitExact untouched), all archived inputs replay clean, bounded local
runs clean (ffmpeg_oracle_decode 300s / 560k execs, panic_free_decode
120s, sei_payload 60s). Also fixes a clippy-1.95 unnecessary_cast drift
in sei_payload.

+++ fuzz/fuzz_targets/sei_payload.rs
@@ -42,8 +42,7 @@ const PAYLOAD_TYPES: &[u32] = &[
     // §D.1.29..§D.1.35 — HDR + 360 family
     137, 142, 144, 147, 148, 149, 150, 151, 154, 155, 156,
     // Annex H §H.13.2.6 — alternative_depth_info
-    181,
-    // §D.1.36..§D.1.38 — transport-layer hints
+    181, // §D.1.36..§D.1.38 — transport-layer hints
     200, 201, 205,
     // Annex F/G/H/I numeric ranges + reserved values — hit the `Unknown`
     // fallback path so the catch-all doesn't go unfuzzed.
@@ -95,7 +94,7 @@ fn context_from_byte(b: u8) -> SeiContext {
         // bounded under fuzzing.
         mvc_view_ref_counts: (0..((b as usize) & 0x03))
             .map(|i| {
-                let n = ((b >> (i % 6)) & 0x03) as u8;
+                let n = (b >> (i % 6)) & 0x03;
                 MvcViewRefCounts {
                     num_anchor_refs_l0: n,
                     num_anchor_refs_l1: (b >> 1) & 0x03,
